requirements.txt
streamlit>=1.35.0
pandas>=2.0,<3
requests
altair<5
pytz
pytest
ruff>=0.11,<1
PyYAML>=6.0
reportlab>=4.0
pypdf>=5.0
